requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
